Add compare token userId and request param id to verify correct token.

+++ api/controller/userController.js
@@ -8,10 +8,11 @@ const SALT_ROUNDS = 12;
 const protectedPath = (req, res, next) => {
   const token = req.headers.authorization;
   if (token) {
-    jwt.verify(token, process.env.JWT_SECRET, (err, decodedToken) => {
+    jwt.verify(token, process.env.JWT_SECRET, (err, decoded) => {
       if (err) {
-        return res.status(401).json({ err });
-      }
+        return res.status(500).json({ message: `Error processing request.` });
+      } 
+      req.userId = decoded.id;
       next();
     })
   } else {
@@ -21,23 +22,24 @@ const protectedPath = (req, res, next) => {
 
 router.all('/*', protectedPath);
 
-router.route('/')
-    .post((req, res) => {
-        const newUser = ({ email, password } = req.body);
-        User.create(newUser)
-            .then(response => res.status(201).json(response))
-            .catch(err => res.status(500).json({ error: err.message }));
-    })
-
 router.route('/:id')
     .get((req, res) => {
         const { id } = req.params;
+        const { userId } = req;
+        if (id !== userId) {
+            return res.status(403).json({ message: `Wrong token.`});
+        }
         User.findById(id)
             .then(response => res.json(response))
             .catch(err => res.status(500).json({ error: err.message }));
     })
     .put((req, res) => {
-        if (req.body.password !== '') {
+        const { id } = req.params;
+        const { userId } = req;
+        if (id !== userId) {
+            return res.status(403).json({ message: `Wrong token.`});
+        }
+        if (req.body.password) {
             bcrypt.hash(req.body.password, SALT_ROUNDS)
             .then(hash => {
                 req.body.password = hash;
@@ -48,17 +50,25 @@ router.route('/:id')
                         res.status(202).json(response)})
                     .catch(err => res.status(500).json({ error: err.message }));
                 })
-            .catch(err => console.log(err));
+            .catch(err => res.status(500).json({ error: err.message }));
         } else {
-            const updateUser = ({ email, firstName, lastName, password } = req.body);
+            const updateUser = ({ email, firstName, lastName } = req.body);
             const { id } = req.params;
-            User.findByIdAndUpdate(id, (updateUser))
-                .then(response => res.status(202).json(response))
+            User.findByIdAndUpdate(id, updateUser)
+                .then(() => {
+                    User.findById(id)
+                        .then(response => res.status(202).json(response))
+                        .catch(err => res.status(500).json({ error: err.message }));
+                })
                 .catch(err => res.status(500).json({ error: err.message }));
         }
     })
     .delete((req, res) => {
         const { id } = req.params;
+        const { userId } = req;
+        if (id !== userId) {
+            return res.status(403).json({ message: `Wrong token.`});
+        }
         User.findByIdAndRemove(id)
             .then(response => res.json(response))
             .catch(err => res.status(500).json({ error: err.message }));
@@ -67,6 +77,10 @@ router.route('/:id')
 router.route('/:id/notes')
     .get((req, res) => {
         const { id } = req.params;
+        const { userId } = req;
+        if (id !== userId) {
+            return res.status(403).json({ message: `Wrong token.`});
+        }
         Note.find({ userId: id })
             .sort('-updated')
             .select('title body')
